SeratoMarkers(2): Cast size to quint32 during serialization
In Qt >= 6.0, the return value of QList::size() and QByteArray::length()
is qsizetype (which may be 8 bytes on 64 bit platforms) instead of a
regular 4 byte int. We add a typecast to quint32 to ensure that the size
is always serialized correctly.

+++ src/track/serato/markers2.cpp
@@ -537,7 +537,7 @@ QByteArray SeratoMarkers2::dumpCommon() const {
         QByteArray entryData = entry->dump();
         stream.writeRawData(entryName.constData(), entryName.length());
         stream << static_cast<quint8>('\x00') // terminating null-byte
-               << entryData.length();
+               << static_cast<quint32>(entryData.length());
         stream.writeRawData(entryData.constData(), entryData.length());
     }
     data.append('\0');
@@ -767,7 +767,7 @@ QByteArray SeratoMarkers2::dumpBase64Encoded() const {
         QByteArray entryData = entry->dump();
         stream.writeRawData(entryName.constData(), entryName.length());
         stream << static_cast<quint8>(0x00) // terminating null-byte
-               << entryData.length();
+               << static_cast<quint32>(entryData.length());
         stream.writeRawData(entryData.constData(), entryData.length());
     }
     innerData.append('\0');
